requirements.txt
Django==5.1.5
djangorestframework==3.15.2
drf-spectacular==0.28.0
black==25.1.0
isort==6.0.0
psycopg==3.2.4
dj-database-url==2.3.0
gunicorn==23.0.0
